fix some casting / boxing issues

+++ v2020-vna-import/src/main/java/org/veo/service/ie/ObjectImportThread.java
@@ -133,11 +133,11 @@ public class ObjectImportThread implements Callable<ObjectImportContext> {
             }
         }
         if (value instanceof Integer) {
-            property.setNumber((long) value);
+            property.setNumber(((Integer) value).longValue());
         }
 
         if (value instanceof Long) {
-            property.setNumber((long) value);
+            property.setNumber((Long) value);
         }
         return property;
     }
